requirements/local.txt
-r production.txt

Werkzeug[watchdog]==3.1.3 # https://github.com/pallets/werkzeug
ipdb==0.13.13  # https://github.com/gotcha/ipdb
psycopg[c]==3.2.7  # https://github.com/psycopg/psycopg

# Testing
# ------------------------------------------------------------------------------
mypy==1.15.0  # https://github.com/python/mypy
django-stubs[compatible-mypy]==5.1.3  # https://github.com/typeddjango/django-stubs
pytest==8.3.5  # https://github.com/pytest-dev/pytest
pytest-sugar==1.0.0  # https://github.com/Teemu/pytest-sugar
djangorestframework-stubs==3.15.3  # https://github.com/typeddjango/djangorestframework-stubs

# Documentation
# ------------------------------------------------------------------------------
sphinx==8.2.3 # pyup: != 8.3.0  # https://github.com/sphinx-doc/sphinx
sphinx-autobuild==2024.10.3 # https://github.com/GaretJax/sphinx-autobuild

# Code quality
# ------------------------------------------------------------------------------
ruff==0.11.8  # https://github.com/astral-sh/ruff
coverage==7.8.0  # https://github.com/nedbat/coveragepy
djlint==1.36.4  # https://github.com/Riverside-Healthcare/djLint
pre-commit==4.2.0  # https://github.com/pre-commit/pre-commit

# Django
# ------------------------------------------------------------------------------
factory-boy==3.3.2  # https://github.com/FactoryBoy/factory_boy

django-debug-toolbar==5.1.0  # https://github.com/jazzband/django-debug-toolbar
django-extensions==4.1  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==3.1.0  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.11.1  # https://github.com/pytest-dev/pytest-django

# consultapp2
# ------------------------------------------------------------------------------
django-filter
djangorestframework-simplejwt
